Change id to be Long

+++ src/main/java/org/oddys/timetracking/dao/ActivityDao.java
@@ -2,5 +2,5 @@ package org.oddys.timetracking.dao;
 
 import org.oddys.timetracking.entity.Activity;
 
-public interface ActivityDao extends Dao<Integer, Activity> {
+public interface ActivityDao extends Dao<Long, Activity> {
 }
