+++ src/main/java/org/oddys/timetracking/dao/RoleDao.java
@@ -2,5 +2,5 @@ package org.oddys.timetracking.dao;
 
 import org.oddys.timetracking.entity.Role;
 
-public interface RoleDao extends Dao<Integer, Role> {
+public interface RoleDao extends Dao<Long, Role> {
 }
