+++ src/main/java/org/oddys/timetracking/entity/Entity.java
@@ -3,20 +3,20 @@ package org.oddys.timetracking.entity;
 import java.io.Serializable;
 
 public abstract class Entity implements Serializable {
-    private Integer id;
+    private Long id;
 
     public Entity() {
     }
 
-    public Entity(Integer id) {
+    public Entity(Long id) {
         this.id = id;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 }
